+++ packages/edit-site/src/components/page-pages/index.js
@@ -22,8 +22,11 @@ import { dateI18n, getDate, getSettings } from '@wordpress/date';
 import Page from '../page';
 import Link from '../routes/link';
 import { DataViews } from '../dataviews';
-import useTrashPostAction from '../actions/trash-post';
-import { postRevisionsAction, viewPostAction } from '../actions/view-actions';
+import {
+	useTrashPostAction,
+	postRevisionsAction,
+	viewPostAction,
+} from '../actions';
 import Media from '../media';
 import DataviewsContext from '../dataviews/context';
 import { DEFAULT_STATUSES } from '../dataviews/provider';
